docs(search): add playground example
