+++ skills/carbon-react/components/search.md
@@ -69,19 +69,6 @@ description: Carbon Search component props and usage examples.
 | warning | string \| boolean \| undefined | No |  | Yes | This prop no longer has any effect. This prop will eventually be removed. |  |  |
 
 ## Examples
-### Default
-
-**Render**
-
-```tsx
-() => {
-  const [value, setValue] = useState("");
-
-  return <Search value={value} onChange={(e) => setValue(e.target.value)} />;
-}
-```
-
-
 ### With Dropdown
 
 **Render**
@@ -163,309 +150,6 @@ description: Carbon Search component props and usage examples.
           setDismissed(true);
         }}
         onClose={() => setDismissed(true)}
-      />
-    </Box>
-  );
-}
-```
-
-
-### With Label and Input Hint
-
-**Render**
-
-```tsx
-() => {
-  const [value, setValue] = useState("Here is some text");
-  return (
-    <Search
-      label="Search"
-      inputHint="Hint text (optional)."
-      onChange={(e) => setValue(e.target.value)}
-      value={value}
-    />
-  );
-}
-```
-
-
-### Sizes
-
-**Render**
-
-```tsx
-() => {
-  const [valueS, setValueS] = useState("");
-  const [valueM, setValueM] = useState("");
-  const [valueL, setValueL] = useState("");
-  return (
-    <Box display="flex" flexDirection="column" gap={3}>
-      <Search
-        label="Small"
-        size="small"
-        onChange={(e) => setValueS(e.target.value)}
-        value={valueS}
-      />
-      <Search
-        label="Medium"
-        size="medium"
-        onChange={(e) => setValueM(e.target.value)}
-        value={valueM}
-      />
-      <Search
-        label="Large"
-        size="large"
-        onChange={(e) => setValueL(e.target.value)}
-        value={valueL}
-      />
-    </Box>
-  );
-}
-```
-
-
-### Sizes with Dropdown
-
-**Render**
-
-```tsx
-() => {
-  const minQueryLength = 2;
-
-  const recentItems = [
-    { value: "Recent term 1", label: "Recent term 1" },
-    { value: "Recent term 2", label: "Recent term 2" },
-    { value: "Recent term 3", label: "Recent term 3" },
-  ];
-
-  const suggestedItems = [
-    {
-      value: "Suggested term 1",
-      label: "Suggested term 1",
-    },
-    {
-      value: "Suggested term 2",
-      label: "Suggested term 2",
-    },
-    {
-      value: "Suggested term 3",
-      label: "Suggested term 3",
-    },
-    {
-      value: "Suggested term 4",
-      label: "Suggested term 4",
-    },
-    {
-      value: "Suggested term 5",
-      label: "Suggested term 5",
-    },
-  ];
-
-  const [valueS, setValueS] = useState("");
-  const [valueM, setValueM] = useState("");
-  const [valueL, setValueL] = useState("");
-  const [dismissedS, setDismissedS] = useState(false);
-  const [dismissedM, setDismissedM] = useState(false);
-  const [dismissedL, setDismissedL] = useState(false);
-
-  const getListData = (val: string) => {
-    const filteredRecent = recentItems.filter((item) =>
-      item.label.toLowerCase().includes(val.toLowerCase()),
-    );
-    const filteredSuggested = suggestedItems.filter((item) =>
-      item.label.toLowerCase().includes(val.toLowerCase()),
-    );
-
-    return [
-      ...(filteredRecent.length > 0
-        ? [
-            {
-              heading: "Recent searches",
-              icon: <Icon type="refresh_clock" />,
-              items: filteredRecent,
-            },
-          ]
-        : []),
-      ...(filteredSuggested.length > 0
-        ? [
-            {
-              heading: "Suggested",
-              icon: <Icon type="search" />,
-              items: filteredSuggested,
-            },
-          ]
-        : []),
-    ];
-  };
-
-  return (
-    <Box display="flex" flexDirection="column" gap={3} height="450px">
-      <Search
-        label="Small"
-        size="small"
-        value={valueS}
-        onChange={(e) => {
-          setValueS(e.target.value);
-          setDismissedS(false);
-        }}
-        onKeyDown={(e) => {
-          if (e.key === "Escape") {
-            setValueS("");
-            setDismissedS(true);
-          }
-        }}
-        onFocus={() => setDismissedS(false)}
-        open={
-          valueS.length >= minQueryLength &&
-          getListData(valueS).length > 0 &&
-          !dismissedS
-        }
-        minQueryLength={minQueryLength}
-        listData={getListData(valueS)}
-        onListItemSelect={(val) => {
-          setValueS(val);
-          setDismissedS(true);
-        }}
-        onClose={() => setDismissedS(true)}
-      />
-      <Search
-        label="Medium"
-        size="medium"
-        value={valueM}
-        onChange={(e) => {
-          setValueM(e.target.value);
-          setDismissedM(false);
-        }}
-        onKeyDown={(e) => {
-          if (e.key === "Escape") {
-            setValueM("");
-            setDismissedM(true);
-          }
-        }}
-        onFocus={() => setDismissedM(false)}
-        open={
-          valueM.length >= minQueryLength &&
-          getListData(valueM).length > 0 &&
-          !dismissedM
-        }
-        minQueryLength={minQueryLength}
-        listData={getListData(valueM)}
-        onListItemSelect={(val) => {
-          setValueM(val);
-          setDismissedM(true);
-        }}
-        onClose={() => setDismissedM(true)}
-      />
-      <Search
-        label="Large"
-        size="large"
-        value={valueL}
-        onChange={(e) => {
-          setValueL(e.target.value);
-          setDismissedL(false);
-        }}
-        onKeyDown={(e) => {
-          if (e.key === "Escape") {
-            setValueL("");
-            setDismissedL(true);
-          }
-        }}
-        onFocus={() => setDismissedL(false)}
-        open={
-          valueL.length >= minQueryLength &&
-          getListData(valueL).length > 0 &&
-          !dismissedL
-        }
-        minQueryLength={minQueryLength}
-        listData={getListData(valueL)}
-        onListItemSelect={(val) => {
-          setValueL(val);
-          setDismissedL(true);
-        }}
-        onClose={() => setDismissedL(true)}
-      />
-    </Box>
-  );
-}
-```
-
-
-### Custom Widths
-
-**Render**
-
-```tsx
-() => {
-  const [value, setValue] = useState("");
-  return (
-    <Box display="flex" flexDirection="column" gap={3}>
-      <Search
-        label="searchWidth"
-        onChange={(e) => setValue(e.target.value)}
-        value={value}
-        inputWidth={25}
-      />
-      <Search
-        label="maxWidth"
-        onChange={(e) => setValue(e.target.value)}
-        value={value}
-        maxWidth="75%"
-      />
-    </Box>
-  );
-}
-```
-
-
-### Inverse
-
-**Render**
-
-```tsx
-() => {
-  const [value, setValue] = useState("Here is some text");
-  return (
-    <Box
-      width="700px"
-      display="flex"
-      flexDirection="column"
-      p={3}
-      backgroundColor="#000000"
-    >
-      <Search
-        label="Inverse"
-        inputHint="Use this prop on darker backgrounds"
-        onChange={(e) => setValue(e.target.value)}
-        value={value}
-        inverse
-      />
-    </Box>
-  );
-}
-```
-
-
-### Label Inline
-
-**Render**
-
-```tsx
-() => {
-  const [value, setValue] = useState("");
-  return (
-    <Box display="flex" flexDirection="column" gap={3}>
-      <Search
-        label="Search"
-        labelInline
-        onChange={(e) => setValue(e.target.value)}
-        value={value}
-      />
-      <Search
-        label="Search with hint"
-        inputHint="Hint text (optional)."
-        labelInline
-        onChange={(e) => setValue(e.target.value)}
-        value={value}
       />
     </Box>
   );
